requirements.txt
streamlit==1.42.0
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10

wikipedia==1.4.0
ddgs==9.5.2

